Check file exists before cleanup

+++ src/test/java/com/blackduck/integration/detect/battery/docker/util/DockerTestDirectories.java
@@ -73,8 +73,10 @@ public class DockerTestDirectories {
 
     public void cleanup() throws IOException {
         File rootTestDir = testDirectory.getParentFile();
-        FileUtils.cleanDirectory(rootTestDir);
-        FileUtils.deleteDirectory(rootTestDir);
+        if (rootTestDir.exists()) {
+            FileUtils.cleanDirectory(rootTestDir);
+            FileUtils.deleteDirectory(rootTestDir);
+        }
     }
 
     public Bind[] getBindings() {
